chatbot-retail/requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-community
faiss-cpu
pandas
PyPDF2
python-docx
pydantic
python-dotenv
openpyxl
